[feat]: [CDS-17424]: Create WinRmExecutorFactoryNG in 930 module and FileBasedWinRmExecutorNG (#31028)
