+++ 400-rest/src/test/java/software/wings/core/winrm/executors/FileBasedWinRmExecutorTest.java
@@ -17,13 +17,13 @@ import static org.mockito.Mockito.doReturn;
 import io.harness.CategoryTest;
 import io.harness.annotations.dev.OwnedBy;
 import io.harness.category.element.UnitTests;
+import io.harness.delegate.task.shell.ConfigFileMetaData;
 import io.harness.delegate.task.winrm.WinRmSessionConfig;
 import io.harness.logging.CommandExecutionStatus;
 import io.harness.logging.LogCallback;
 import io.harness.rule.Owner;
 
 import software.wings.beans.ConfigFile;
-import software.wings.beans.command.CopyConfigCommandUnit;
 import software.wings.core.ssh.executors.FileBasedWinRmExecutor;
 import software.wings.delegatetasks.DelegateFileManager;
 
@@ -42,15 +42,14 @@ public class FileBasedWinRmExecutorTest extends CategoryTest {
   @Mock WinRmSessionConfig config;
   @Mock DelegateFileManager delegateFileManager;
   private ConfigFile configFile = ConfigFile.builder().encrypted(false).entityId("TEST_ID").build();
-  private CopyConfigCommandUnit.ConfigFileMetaData configFileMetaData =
-      CopyConfigCommandUnit.ConfigFileMetaData.builder()
-          .destinationDirectoryPath("TEST_PATH")
-          .fileId(configFile.getUuid())
-          .filename("TEST_FILE_NAME")
-          .length(configFile.getSize())
-          .encrypted(configFile.isEncrypted())
-          .activityId("TEST_ACTIVITY_ID")
-          .build();
+  private ConfigFileMetaData configFileMetaData = ConfigFileMetaData.builder()
+                                                      .destinationDirectoryPath("TEST_PATH")
+                                                      .fileId(configFile.getUuid())
+                                                      .filename("TEST_FILE_NAME")
+                                                      .length(configFile.getSize())
+                                                      .encrypted(configFile.isEncrypted())
+                                                      .activityId("TEST_ACTIVITY_ID")
+                                                      .build();
 
   private FileBasedWinRmExecutor spyFileBasedWinRmExecutor;
